refactor : Config Property들 패키지 변경 (#32)
