+++ cmd/config/dbconfig.go
@@ -1,7 +1,8 @@
-package util
+package config
 
 import (
 	"fmt"
+	"guestbook/common/util"
 )
 
 type DbConfig struct {
@@ -19,28 +20,31 @@ type DbConfig struct {
 }
 
 const (
-	FileName        = "common\\property\\db-property.yml"
-	DataSource      = "%s:%s@/%s?parseTime=true"
-	DatabaseAddress = "%s:%s"
+	dbPropertyPath        = "common\\property\\db-property.yml"
+	dataSourceFormat      = "%s:%s@/%s?parseTime=true"
+	databaseAddressFormat = "%s:%s"
 )
 
-func GetDbConfig() (DbConfig, error) {
+func getDbConfig() DbConfig {
 	config := DbConfig{}
-	err := fetchYAML(FileName, &config)
-	return config, err
+	err := util.FetchYAML(dbPropertyPath, &config)
+	if err != nil {
+		panic(err)
+	}
+	return config
 }
 
 func (config DbConfig) DataSource() string {
 	user := config.Db.Mysql.User
 	password := config.Db.Mysql.Password
 	dbname := config.Db.Mysql.DbName
-	return fmt.Sprintf(DataSource, user, password, dbname)
+	return fmt.Sprintf(dataSourceFormat, user, password, dbname)
 }
 
 func (config DbConfig) DatabaseAddress() string {
 	host := config.Db.Mysql.Host
 	port := config.Db.Mysql.Port
-	return fmt.Sprintf(DatabaseAddress, host, port)
+	return fmt.Sprintf(databaseAddressFormat, host, port)
 }
 
 func (config DbConfig) Driver() string {
